Add script to convert mlox rules to LOOT metadata
The script parses the mlox rules into an abstract representation. There
are two other parsers that I know of:

- mlox's is also written in Python, but mixes together parsing and
  evaluation, so isn't suitable for this use case
- plox's seems like a better starting point, but isn't directly usable
  in its current form

This script doesn't attempt to match either parser's behaviour exactly
(e.g. mlox treats whitespace as optional in more places), but instead
targets the syntax that is actually in use in the rules files present in
the https://github.com/mlox/mlox and
https://github.com/DanaePlays/mlox-rules repositories.

While the parser can handle everything in those files, not everything
supported by mlox is supported by LOOT, and the remaining differences
are intentional. See the comment at the top of the script for more
details on the differences and how they're handled.