Add scripts to download and read Morrowind Modding History mods
The script takes the CSV export of the index of mods that were hosted on
Morrowind Modding History and Great House Fliggerty before they went
down, as downloaded from <https://modlist.altervista.org/mmh/>. That CSV
is included in this commit, and results in 61 GB of downloaded files.

This commit also includes a script that reads the index CSV and
downloads directory and outputs a TSV file of all the plugin files in
the downloaded files, looking inside archives. The output TSV file is
also included in this commit.

Both scripts include a few hacks to workaround issues affecting a small
minority of mods:

- The mod IDs 2770, 2771 and 2772 are three separate uploads but form
  the three parts of a multipart 7z archive, so their downloaded files
  are named to allow the archive to be read by 7-zip.
- There are a few downloads that are just bare plugin files: these are
  written into zip files so that the plugin filenames can be preserved,
  since the download file name needs to be the mod ID.
- Some download URLs use tinyurl: these have a hardcoded mapping to the
  URLs that tinyurl points to, in case the tinyurl URLs stop working.
- Nexus Mods URLs are skipped as they're already covered by the
  fetch_nexus_plugins_index.py script.

The download script also handles the many cases where mod IDs point to
downloads that are no longer available, by skipping a mod entirely if
its ID appears in a hardcoded whitelist of known invalid downloads.